+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Limited amount of shots per second for each player
 * Players keep firing while trigger is pressed
 * Different colors for the two players
+* Players can also fire using the right trigger
 
 ## [0.1.1] - 2019-12-21
 ### Fixed